backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pgvector==0.3.5
python-docx==1.1.2
python-multipart==0.0.9
anthropic>=0.18.0
openai>=1.0.0  # 百炼 OpenAI 兼容格式
httpx>=0.27.0
sentence-transformers>=3.0.0
torch>=2.0.0
pydantic==2.10.0
pydantic-settings==2.6.0
alembic==1.13.0
